\name{DiscreteLaplace-package}
\alias{DiscreteLaplace-package}
\alias{DiscreteLaplace}
\docType{package}
\title{
Discrete Laplace distribution
}
\description{
Density, distribution function, quantile function and random generation for the skew discrete Laplace distribution
}
\details{
\tabular{ll}{
Package: \tab DiscreteLaplace\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-01-29\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Alessandro Barbiero, Riccardo Inchingolo
Maintainer: Alessandro Barbiero <alessandro.barbiero@unimi.it>
}
\references{
Tomasz J. Kozubowski, Seidu Inusah (2006) A skew Laplace distribution on integers, \emph{AISM}, 58: 555-571
}
\keyword{package}

