\name{DiscreteLaplace-package}
\alias{DiscreteLaplace-package}
\alias{DiscreteLaplace}
\docType{package}
\title{
Discrete Laplace Distributions
}
\description{
Probability mass function, distribution function, quantile function, random generation and sample estimation for two discrete skew Laplace distributions on integers. The skew discrete Laplace distributions here considered are that proposed by Kozubowski and Inusah (2006), henceforth referred to as DSL, and the alternative one proposed by Barbiero (2014), henceforth ADSL.
}
\details{
\tabular{ll}{
Package: \tab DiscreteLaplace\cr
Type: \tab Package\cr
Version: \tab 1.1.1\cr
Date: \tab 2016-04-29\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Alessandro Barbiero, Riccardo Inchingolo

Maintainer: Alessandro Barbiero <alessandro.barbiero@unimi.it>
}
\references{
T. J. Kozubowski, S. Inusah (2006) A skew Laplace distribution on integers, \emph{Annals of the Institute of Statistical Mathematics}, 58: 555-571, \url{http://dx.doi.org/10.1007/s10463-005-0029-1}

A. Barbiero (2014) An alternative discrete Laplace distribution, \emph{Statistical Methodology}, 16: 47-67, \url{http://dx.doi.org/10.1016/j.stamet.2013.07.002}
}
\keyword{package}